Fix 1.4 compatibility
git-svn-id: file:///home/toad/git-migration/temprepository/trunk/apps/Thaw@10665 67a373e5-eb02-0410-a15c-ee090a768436

+++ src/thaw/plugins/index/SearchResult.java
@@ -65,7 +65,7 @@ public class SearchResult extends Observable implements FileAndLinkList {
 
 		for (Iterator it = indexIds.iterator();
 		     it.hasNext(); i++) {
-			st.setInt(i, (Integer)it.next());
+			st.setInt(i, ((Integer)it.next()).intValue());
 		}
 
 		for (int j = 0 ; j < searchPatterns.length; j++) {
